requirements.txt
torch
torchvision
torch-geometric
numpy
pandas
scipy
pyyaml
matplotlib
seaborn
tqdm
scikit-learn
tabulate
tensorboard
